+++ lpf/converters/schnakenbergconverter.py
@@ -20,7 +20,7 @@ class SchnakenbergModel(Converter):
 
     def to_params(self, dv, params=None):
         if params is None:
-            params = np.zeros((1, 8), dtype=np.float64)
+            params = np.zeros((1, 8), dtype=np.float32)
 
         params[0, 0] = 10 ** dv[0, 0]  # Du
         params[0, 1] = 10 ** dv[0, 1]  # Dv
@@ -33,7 +33,7 @@ class SchnakenbergModel(Converter):
 
     def to_init_states(self, dv, init_states=None):
         if init_states is None:
-            init_states = np.zeros((1, 2), dtype=np.float64)
+            init_states = np.zeros((1, 2), dtype=np.float32)
 
         init_states[0, 0] = 10 ** dv[0, 6]  # u0
         init_states[0, 1] = 10 ** dv[0, 7]  # v0
